LRAC-10637 move and rename
